+++ ixnetwork_restpy/testplatform/sessions/ixnetwork/topology/cfmbridge_0d1f83a1e9fee8ee7d444b2a33a0e03b.py
@@ -727,6 +727,26 @@ class CfmBridge(Base):
         for item in kwargs.items(): payload[item[0]] = item[1]
         return self._execute('getCfmLoopbackDbLearnedInformation', payload=payload, response_object=None)
 
+    def GetCfmSLMDbLearnedInformation(self, *args, **kwargs):
+        """Executes the getCfmSLMDbLearnedInformation operation on the server.
+
+        Get Learned SLM Information
+
+        getCfmSLMDbLearnedInformation(Arg2=list)list
+        --------------------------------------------
+        - Arg2 (list(number)): List of indices into the protocol plugin.An empty list indicates all instances in the plugin.
+        - Returns list(str): Please provide a proper description here.
+
+        Raises
+        ------
+        - NotFoundError: The requested resource does not exist on the server
+        - ServerError: The server has encountered an uncategorized error condition
+        """
+        payload = { "Arg1": self.href }
+        for i in range(len(args)): payload['Arg%s' % (i + 2)] = args[i]
+        for item in kwargs.items(): payload[item[0]] = item[1]
+        return self._execute('getCfmSLMDbLearnedInformation', payload=payload, response_object=None)
+
     def GetCfmTSTDbLearnedInformation(self, *args, **kwargs):
         """Executes the getCfmTSTDbLearnedInformation operation on the server.
 
@@ -1051,6 +1071,31 @@ class CfmBridge(Base):
         for i in range(len(args)): payload['Arg%s' % (i + 2)] = args[i]
         for item in kwargs.items(): payload[item[0]] = item[1]
         return self._execute('getPeriodicLTLearnedInformation', payload=payload, response_object=None)
+
+    def GetSLMLearnedInfo(self, *args, **kwargs):
+        """Executes the getSLMLearnedInfo operation on the server.
+
+        Get SLM Learned Info
+
+        The IxNetwork model allows for multiple method Signatures with the same name while python does not.
+
+        getSLMLearnedInfo(SessionIndices=list)
+        --------------------------------------
+        - SessionIndices (list(number)): This parameter requires an array of session numbers 1 2 3
+
+        getSLMLearnedInfo(SessionIndices=string)
+        ----------------------------------------
+        - SessionIndices (str): This parameter requires a string of session numbers 1-4;6;7-12
+
+        Raises
+        ------
+        - NotFoundError: The requested resource does not exist on the server
+        - ServerError: The server has encountered an uncategorized error condition
+        """
+        payload = { "Arg1": self }
+        for i in range(len(args)): payload['Arg%s' % (i + 2)] = args[i]
+        for item in kwargs.items(): payload[item[0]] = item[1]
+        return self._execute('getSLMLearnedInfo', payload=payload, response_object=None)
 
     def GetTSTLearnedInfo(self, *args, **kwargs):
         """Executes the getTSTLearnedInfo operation on the server.