+++ ixnetwork_restpy/testplatform/sessions/ixnetwork/topology/cfmsimulatedmp_d0096b7abd3340f4c8bdf5e2ba2578fb.py
@@ -1745,6 +1745,31 @@ class CfmSimulatedMp(Base):
         for item in kwargs.items(): payload[item[0]] = item[1]
         return self._execute('startCcmSimulatedMp', payload=payload, response_object=None)
 
+    def StartSlm(self, *args, **kwargs):
+        """Executes the startSlm operation on the server.
+
+        Start SLM
+
+        The IxNetwork model allows for multiple method Signatures with the same name while python does not.
+
+        startSlm(SessionIndices=list)
+        -----------------------------
+        - SessionIndices (list(number)): This parameter requires an array of session numbers 1 2 3
+
+        startSlm(SessionIndices=string)
+        -------------------------------
+        - SessionIndices (str): This parameter requires a string of session numbers 1-4;6;7-12
+
+        Raises
+        ------
+        - NotFoundError: The requested resource does not exist on the server
+        - ServerError: The server has encountered an uncategorized error condition
+        """
+        payload = { "Arg1": self }
+        for i in range(len(args)): payload['Arg%s' % (i + 2)] = args[i]
+        for item in kwargs.items(): payload[item[0]] = item[1]
+        return self._execute('startSlm', payload=payload, response_object=None)
+
     def Stop(self):
         """Executes the stop operation on the server.
 
@@ -1802,3 +1827,28 @@ class CfmSimulatedMp(Base):
         for i in range(len(args)): payload['Arg%s' % (i + 2)] = args[i]
         for item in kwargs.items(): payload[item[0]] = item[1]
         return self._execute('stopCcmSimulatedMp', payload=payload, response_object=None)
+
+    def StopSlm(self, *args, **kwargs):
+        """Executes the stopSlm operation on the server.
+
+        Stop SLM
+
+        The IxNetwork model allows for multiple method Signatures with the same name while python does not.
+
+        stopSlm(SessionIndices=list)
+        ----------------------------
+        - SessionIndices (list(number)): This parameter requires an array of session numbers 1 2 3
+
+        stopSlm(SessionIndices=string)
+        ------------------------------
+        - SessionIndices (str): This parameter requires a string of session numbers 1-4;6;7-12
+
+        Raises
+        ------
+        - NotFoundError: The requested resource does not exist on the server
+        - ServerError: The server has encountered an uncategorized error condition
+        """
+        payload = { "Arg1": self }
+        for i in range(len(args)): payload['Arg%s' % (i + 2)] = args[i]
+        for item in kwargs.items(): payload[item[0]] = item[1]
+        return self._execute('stopSlm', payload=payload, response_object=None)