+++ ixnetwork_restpy/testplatform/sessions/ixnetwork/topology/startccmemulatedmpparams_62b1a2be6189ae1e1aa5f0cee8254081.py
@@ -39,7 +39,7 @@ class StartCcmEmulatedMpParams(Base):
     def StartCcmEmulatedMp(self):
         """Executes the startCcmEmulatedMp operation on the server.
 
-        Start CCM PDU Transmission
+        Start CCM PDU Transmission.
 
         Raises
         ------