+++ ixnetwork_restpy/testplatform/sessions/ixnetwork/topology/stopccmemulatedmpparams_a24dfdb02ad9d3f95459876f02f9eff0.py
@@ -39,7 +39,7 @@ class StopCcmEmulatedMpParams(Base):
     def StopCcmEmulatedMp(self):
         """Executes the stopCcmEmulatedMp operation on the server.
 
-        Stop CCM PDU Transmission
+        Stop CCM PDU Transmission.
 
         Raises
         ------